+++ pysui/sui/sui_txn/async_transaction.py
@@ -597,16 +597,9 @@ class SuiTransactionAsync(_SuiTransactionBase):
             arguments = (
                 arguments if isinstance(arguments, list) else arguments.array
             )
-            arguments = await self._resolve_arguments(arguments)
-            for index, arg in enumerate(arguments):
-                parm = parameters[index]
-                if (
-                    hasattr(parm, "is_mutable")
-                    and parm.is_mutable
-                    and isinstance(arg, tuple)
-                ):
-                    r_arg: bcs.ObjectArg = arg[1]
-                    r_arg.value.Mutable = True
+            arguments = self._receiving_feature(
+                await self._resolve_arguments(arguments), parameters
+            )
         else:
             arguments = []
         # Standardize the type_arguments to list
@@ -663,16 +656,7 @@ class SuiTransactionAsync(_SuiTransactionBase):
             res_count,
         ) = await self._move_call_target_cache(target)
         if arguments:
-            for index, arg in enumerate(arguments):
-                parm = parameters[index]
-                if (
-                    hasattr(parm, "is_mutable")
-                    and parm.is_mutable
-                    and isinstance(arg, tuple)
-                ):
-                    r_arg: bcs.ObjectArg = arg[1]
-                    r_arg.value.Mutable = True
-
+            arguments = self._receiving_feature(arguments, parameters)
         type_arguments = (
             type_arguments if isinstance(type_arguments, list) else []
         )