+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.37.0] - 2023-10-02
+
+### Added
+
+- [enhancement](https://github.com/FrankC01/pysui/issues/165) Receiving object call argument
+- [enhancement](https://github.com/FrankC01/pysui/issues/166) Start using protocol config for feature constraints. First application is for Receiving
+
+### Fixed
+
+### Changed
+
+### Removed
+
 ## [0.36.1] - Unpublished
 
 ### Added